refactor: modularize test files by splitting app_test.go into domain-specific test suites
