+++ .github/codecov.yml
@@ -14,3 +14,5 @@ ignore:
   - "frontend/wailsjs/**/*"
   - "build/**/*"
   - "main.go"              # wails.Run など自動テストが困難なエントリーポイント
+  - "app_dialogs.go"       # ダイアログ操作などテスト困難なUI
+  - "app_fonts.go"         # システム環境依存のフォント取得
